requirements.txt
discord.py~=2.5.0
mariadb~=1.1.12
dotenv~=0.9.9